requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
aiohttp-requests==0.1.3
async-timeout==3.0.1
attrs==20.2.0
cchardet==2.1.6
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
discord.py==1.4.1
idna==2.10
multidict==4.7.6
pycares==3.1.1
pycparser==2.20
requests==2.24.0
urllib3==1.25.10
yarl==1.5.1